Day 13 part 2 done

+++ day13/src/part1.rs
@@ -1,7 +1,7 @@
 #[derive(Debug)]
 pub(crate) struct Vector {
-    horizontal: i32,
-    vertical: i32
+    horizontal: i64,
+    vertical: i64
 }
 
 impl Vector {
@@ -11,15 +11,15 @@ impl Vector {
         Some(Self::new(nums.0, nums.1))
     }
 
-    fn new(horizontal: i32, vertical: i32) -> Vector {
+    fn new(horizontal: i64, vertical: i64) -> Vector {
         Self { horizontal, vertical }
     }
 }
 
 #[derive(Debug)]
 pub(crate) struct Coordinate {
-    horizontal: i32,
-    vertical: i32,
+    horizontal: i64,
+    vertical: i64,
 }
 
 impl Coordinate {
@@ -28,15 +28,15 @@ impl Coordinate {
         Some(Self::new(nums.0, nums.1))
     }
 
-    pub(crate) fn new(horizontal: i32, vertical: i32) -> Coordinate {
+    pub(crate) fn new(horizontal: i64, vertical: i64) -> Coordinate {
         Coordinate { horizontal, vertical }
     }
 }
 
 #[derive(Debug)]
 pub(crate) struct ClawMachine {
-    price_a: i32,
-    price_b: i32,
+    price_a: i64,
+    price_b: i64,
     button_a: Vector,
     button_b: Vector,
     target: Coordinate
@@ -56,7 +56,7 @@ impl ClawMachine {
         Self { button_a: ba, button_b: bb, target, price_a: 3, price_b: 1 }
     }
 
-    pub(crate) fn least_tokens(&self) -> Option<i32> {
+    pub(crate) fn least_tokens(&self) -> Option<i64> {
         // normalize around A
         // target.x = pressA * A.x + pressB * B.x
         // target.y = pressA * A.y + pressB * B.y
@@ -77,12 +77,12 @@ impl ClawMachine {
 
         // println!("all right, calculating values for claw machine\n{:?}", self);
         let press_a = (self.button_b.horizontal * self.target.vertical
-            - self.button_b.vertical * self.target.horizontal) as f32
-        / (self.button_a.vertical * self.button_b.horizontal - self.button_a.horizontal * self.button_b.vertical) as f32;
+            - self.button_b.vertical * self.target.horizontal) as f64
+        / (self.button_a.vertical * self.button_b.horizontal - self.button_a.horizontal * self.button_b.vertical) as f64;
 
         // println!("press a: {}", press_a);
 
-        if press_a.fract() != 0f32 {
+        if press_a.fract() != 0f64 {
             // println!("-- press a fraction is : {}", press_a.fract());
             return None
         }
@@ -92,24 +92,24 @@ impl ClawMachine {
         // target.x - pressA * A.x = pressB * B.x
         // (target.x - pressA * A.x) / B.x = pressB
 
-        let press_b = (self.target.horizontal as f32 - press_a * self.button_a.horizontal as f32) / self.button_b.horizontal as f32;
+        let press_b = (self.target.horizontal as f64 - press_a * self.button_a.horizontal as f64) / self.button_b.horizontal as f64;
 
         // println!("press b: {}", press_b);
-        if press_b.fract() != 0f32 {
+        if press_b.fract() != 0f64 {
             // println!("-- press b fraction is : {}", press_b.fract());
             return None
         }
 
-        Some((press_a * self.price_a as f32 + press_b * self.price_b as f32) as i32)
+        Some((press_a * self.price_a as f64 + press_b * self.price_b as f64) as i64)
     }
 }
 
-pub(crate) fn str_to_nums(input: &str) -> Option<(i32, i32)> {
+pub(crate) fn str_to_nums(input: &str) -> Option<(i64, i64)> {
     let nums = input
         .trim()
         .split(", ")
-        .map(|x| x[2..].parse::<i32>().unwrap())
-        .collect::<Vec<i32>>();
+        .map(|x| x[2..].parse::<i64>().unwrap())
+        .collect::<Vec<i64>>();
 
     if nums.len() != 2 {
         return None
@@ -120,7 +120,7 @@ pub(crate) fn str_to_nums(input: &str) -> Option<(i32, i32)> {
 
 
 
-pub(crate) fn solve(input: &str) -> i32 {
+pub(crate) fn solve(input: &str) -> i64 {
     let machines:Vec<ClawMachine> = input
         .trim()
         .split("\n\n")
